refactor(step-wizard): add types folder
