+++ packages/step-wizard/index.d.ts
@@ -1,9 +1,4 @@
-import Wizard from './typings/Wizard';
-import WizardStep from './typings/WizardStep';
-import WizardStepBadge from './typings/WizardStepBadge';
-import WizardStepTitle from './typings/WizardStepTitle';
-
-export default Wizard;
-
-export { WizardStep, WizardStepBadge, WizardStepTitle };
-
+export { default } from './types/Wizard';
+export { default as WizardStep } from './types/WizardStep';
+export { default as WizardStepBadge } from './types/WizardStepBadge';
+export { default as WizardStepTitle } from './types/WizardStepTitle';
