+++ packages/step-wizard/package.json
@@ -4,6 +4,7 @@
   "author": "Tyson Warner <tyson.warner@availity.com>",
   "description": "Availity step wizard",
   "main": "index.js",
+  "types": "index.d.ts",
   "keywords": [
     "react",
     "availity",
@@ -22,7 +23,6 @@
     "access": "public"
   },
   "devDependencies": {
-    "prop-types": "^15.7.2",
     "react": "^17.0.2",
     "react-dom": "^17.0.2"
   },
@@ -30,6 +30,7 @@
     "react": ">=16.3.0"
   },
   "dependencies": {
+    "prop-types": "^15.7.2",
     "classnames": "^2.3.1"
   }
 }
