+++ packages/step-wizard/types/Wizard.d.ts
@@ -1,0 +1,11 @@
+export interface WizardProps {
+  bar?: boolean;
+  stacked?: boolean;
+  progress?: boolean;
+  tag?: React.ReactType | string;
+  className?: string;
+}
+
+declare const Wizard: React.FC<WizardProps>;
+
+export default Wizard;
