+++ packages/step-wizard/types/WizardStepBadge.d.ts
@@ -1,0 +1,8 @@
+export interface WizardStepBadgeProps {
+  className?: string;
+  tag?: React.ReactType | string;
+}
+
+declare const WizardStepBadge: React.FC<WizardStepBadgeProps>;
+
+export default WizardStepBadge;
