+++ packages/step-wizard/types/WizardStepTitle.d.ts
@@ -1,0 +1,8 @@
+export interface WizardStepTitleProps {
+  className?: string;
+  tag?: React.ReactType | string;
+}
+
+declare const WizardStepTitle: React.FC<WizardStepTitleProps>;
+
+export default WizardStepTitle;
