--- packages/step-wizard/typings/WizardStep.d.ts
@@ -1,15 +1,0 @@
-export interface WizardStepProps {
-    complete?: boolean;
-    active?: boolean;
-    disabled?: boolean;
-    clickable?: boolean;
-    href?: string;
-    children?: React.ReactType;
-    tag?: React.ReactType | string;
-    className?: string;
-}
-
-declare const WizardStep: React.FunctionComponent<WizardStepProps>;
-
-export default WizardStep;
-
