--- packages/step-wizard/typings/WizardStepBadge.d.ts
@@ -1,9 +1,0 @@
-export interface WizardStepBadgeProps {
-    className?: string;
-    children?: React.ReactType;
-    tag?: React.ReactType | string;
-}
-
-declare const WizardStepBadge: React.FunctionComponent<WizardStepBadgeProps>;
-
-export default WizardStepBadge;
