--- packages/step-wizard/typings/WizardStepTitle.d.ts
@@ -1,9 +1,0 @@
-export interface WizardStepTitleProps {
-    className?: string;
-    children?: React.ReactType;
-    tag?: React.ReactType | string;
-}
-
-declare const WizardStepTitle: React.FunctionComponent<WizardStepTitleProps>;
-
-export default WizardStepTitle;
